fix/AuzreOpenAI embeddings and LLM deployment name
solves "The API deployment for this resource does not exist" for LLM and
embedding models deployed in Azure OpenAI by deployment name supported
in tmc/langchaingo#253

We can't Validate OpenAI LLM model names from hard-coded list in Azure
because the model name parameter in API request is a deployment name,
and while Microsoft advises us to use the model name as deployment name,
we did not listen, and I didn't want to coordinate redeploying with a
different name on a Friday.

This also permits use of customized models that can be deployed in Azure
side-by-side base models as added benefit so I think it was worthwhile.