build(golangci): Update linters and output settings
Add new linters and modify existing ones in the "golangci.yaml" file.
Update the max number of issues per linter and the max number of the same issues.
Additionally, make changes to the settings of the "gocritic" linter.
In addition fix the corresponding go files regarding the changes from linter settings.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>